Check if mountpoint is a directory before deleting

+++ backupctl/backupctl.py
@@ -301,7 +301,8 @@ def remove(hist, pool, customer, vault=None):
     else:
         fs = os.path.join(pool, customer)
     zfs.remove_filesystem(fs)
-    os.rmdir(fs)
+    if os.path.isdir(fs):
+        os.rmdir(fs)
     hist.add(customer, "remove", vault)
